projects/ai-inference/requirements.txt
msgpack
numpy
tensorflow==2.1.0
requests
paho-mqtt